cockpit-actions: Concentrate types on single file
